Refactor CI for pull requests
This is the first step in refactoring the CI in this repo.
One of the goals is to avoid re-building the snap on testflinger
devices by publishing the snap to latest/edge where the testflinger
job can install from. This also adds some additional checks on the
snapcraft.yaml and pushes a tag whose name corresponds to the snap
revision for improved tracking.