Refactor CI for pull requests (#18)
* Refactor CI for pull requests

This is the first step in refactoring the CI in this repo.
One of the goals is to avoid re-building the snap on testflinger
devices by publishing the snap to latest/edge where the testflinger
job can install from. This also adds some additional checks on the
snapcraft.yaml and pushes a tag whose name corresponds to the snap
revision for improved tracking.

* Add plug and slot declarations for snap review tool

* Remove stray double quote

* Comment release step due to missing override in the Store

* Test pushing new tag

* Use hyphen instead of slash to fix reference error

* Simplify and refactor integration testing